Title Screen Changes
Trying to set up a makeshift title screen and game loop but I only got
as far as canvases, buttons and canvas scaling with the camera etc.
Sandbox is still practically empty save for a rudimentary VR camera and
terrain. definitely want to set up an external view for the monitor and
I'm thinking a UI built into the player's arms akin to HLVR would be
awesome!